Add mocha unit tests for logic/init_transaction
