+++ framework/test/mocha/unit/modules/chain/logic/test_data/index.js
@@ -1,0 +1,25 @@
+/*
+ * Copyright © 2018 Lisk Foundation
+ *
+ * See the LICENSE file at the top-level directory of this distribution
+ * for licensing information.
+ *
+ * Unless otherwise agreed in a custom licensing agreement with the Lisk Foundation,
+ * no part of this software, including this file, may be copied, modified,
+ * propagated, or distributed except according to the terms contained in the
+ * LICENSE file.
+ *
+ * Removal or modification of this copyright notice is prohibited.
+ */
+
+'use strict';
+
+const dapp = require('./dapp');
+const multisignature = require('./multisignature');
+const out_transfer = require('./out_transfer');
+
+module.exports = {
+	dapp,
+	multisignature,
+	out_transfer,
+};
